fix autorizaçoes

+++ automanager/src/main/java/com/autobots/automanager/controles/EmailControle.java
@@ -2,10 +2,14 @@ package com.autobots.automanager.controles;
 
 import com.autobots.automanager.dtos.emails.CadastroEmailDTO;
 import com.autobots.automanager.entidades.Email;
+import com.autobots.automanager.entidades.Endereco;
+import com.autobots.automanager.entidades.Usuario;
+import com.autobots.automanager.modelos.Perfil;
 import com.autobots.automanager.modelos.adicionadorLinks.AdicionadorLinkEmail;
 import com.autobots.automanager.modelos.atualizador.EmailAtualizador;
 import com.autobots.automanager.repositorios.EmailRepositorio;
 import com.autobots.automanager.repositorios.UsuarioRepositorio;
+import com.autobots.automanager.service.AuthService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -26,6 +30,8 @@ public class EmailControle {
 	private AdicionadorLinkEmail adicionadorLinkEmail;
 	@Autowired
 	private EmailAtualizador atualizarEmail;
+	@Autowired
+	private AuthService authService;
 
 	@PreAuthorize("hasAnyRole('ADMIN', 'GERENTE', 'VENDEDOR')")
 	@GetMapping("get/unique/{id}")
@@ -35,12 +41,16 @@ public class EmailControle {
 		if(email == null){
 			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Email não encontrado");
 		} else {
+
+			var verificacao = verificarPermissao(authService.obterUsuarioLogado(), email);
+			if(verificacao != null) return verificacao;
+
 			adicionadorLinkEmail.adicionarLink(email);
 			return ResponseEntity.status(HttpStatus.OK).body(email);
 		}
 	}
 
-	@PreAuthorize("hasAnyRole('ADMIN', 'GERENTE', 'VENDEDOR')")
+	@PreAuthorize("hasAnyRole('ADMIN')")
 	@GetMapping("get/all")
 	public ResponseEntity<List<Email>> obterEmails() {
 		List<Email> emails = emailRepositorio.findAll();
@@ -57,12 +67,21 @@ public class EmailControle {
 	public ResponseEntity<?> cadastrarEmail(@RequestBody CadastroEmailDTO data) {
 		var cliente = usuarioRepositorio.findById(data.getClienteId()).orElseThrow(
 				() -> new RuntimeException("Cliente não encontrado"));
-			var email = new Email(data.getEmail());
-			cliente.getEmails().add(email);
-			emailRepositorio.save(email);
-			usuarioRepositorio.save(cliente);
-			adicionadorLinkEmail.adicionarLink(email);
-			return ResponseEntity.status(HttpStatus.CREATED).body(email);
+
+		var email = new Email(data.getEmail());
+
+		var validacao = verificarPermissao(authService.obterUsuarioLogado(), email);
+		if(validacao != null) return validacao;
+
+		// Adiciona o email ao cliente
+		email.setUsuario(cliente);
+		cliente.getEmails().add(email);
+
+		emailRepositorio.save(email);
+		usuarioRepositorio.save(cliente);
+		adicionadorLinkEmail.adicionarLink(email);
+
+		return ResponseEntity.status(HttpStatus.CREATED).body(email);
 	}
 
 	@PreAuthorize("hasAnyRole('ADMIN', 'GERENTE', 'VENDEDOR')")
@@ -70,6 +89,9 @@ public class EmailControle {
 	public ResponseEntity<?> atualizarEmail(@RequestBody Email email) {
 		var emailDb = emailRepositorio.findById(email.getId()).orElseThrow(
 				() -> new RuntimeException("Email não encontrado"));
+
+		var validacao = verificarPermissao(authService.obterUsuarioLogado(), emailDb);
+		if(validacao != null) return validacao;
 
 		atualizarEmail.atualizar(emailDb, email);
 		emailRepositorio.save(emailDb);
@@ -81,10 +103,34 @@ public class EmailControle {
 	@DeleteMapping("/excluir/{id}")
 	public ResponseEntity<?> excluirEmail(@PathVariable Long id) {
 		try {
+			var email = emailRepositorio.findById(id).orElseThrow(() -> new RuntimeException("Email não encontrado"));
+
+			var validacao = verificarPermissao(authService.obterUsuarioLogado(), email);
+			if(validacao != null) return validacao;
+
 			emailRepositorio.deleteById(id);
 			return ResponseEntity.status(HttpStatus.OK).body("Email excluído com sucesso");
 		} catch(Exception ex) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Erro ao excluir email");
 		}
+	}
+
+	public ResponseEntity<?> verificarPermissao(Usuario autor, Email email) {
+		if(email.getUsuario() != null){
+			var perfis = email.getUsuario().getPerfis();
+
+			if (perfis.contains(Perfil.ROLE_ADMIN) && !autor.getPerfis().contains(Perfil.ROLE_ADMIN)) {
+				return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Você não tem permissão para acessar este recurso");
+			}
+
+			if (perfis.contains(Perfil.ROLE_GERENTE) && !autor.getPerfis().contains(Perfil.ROLE_GERENTE)) {
+				return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Você não tem permissão para acessar este recurso");
+			}
+
+			if (perfis.contains(Perfil.ROLE_VENDEDOR) && !autor.getPerfis().contains(Perfil.ROLE_GERENTE)) {
+				return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Você não tem permissão para acessar este recurso");
+			}
+		}
+		return null;
 	}
 }
